fixing same bug again

+++ MyPro+UpdThingIDF/main/main.cpp
@@ -116,11 +116,10 @@ void setup() {
 void loop() {
   int sliceSize = 500;
  
-  uint32_t touch_filter_value;
+  uint16_t touch_filter_value;
   #if TOUCH_FILTER_MODE_EN
    // printf("Touch Sensor filter mode read, the output format is: \nTouchpad num:[raw data, filtered data]\n\n");
     // If open the filter mode, please use this API to get the touch pad count.
-    
     touch_pad_read_filtered(TOUCH_PAD_NUM0, &touch_filter_value);
     //printf("T%d:[%4d,%4d] ", TOUCH_PAD_NUM0, touch_value, touch_filter_value);
   #endif
